data_manager/requirements.txt
django-nested-admin
